Name postpone/advance menu items more accurately
This adds a couple of words to the menu items in order to make them more accurate.

Of course, the postpone and advance functions are explained well in the add-on description on Ankiweb and in the dialog that appears once someone clicks on the menu item.

This renaming would just make the menu items a bit more self-explanatory, in my (perhaps too pedantic) opinion. 

It follows the word order of the names of the gear actions and intends to make it obvious that the menu item does exactly the same but to all decks at once.

I didn't rename “Reschedule all cards” because unlike Postpone/Advance, it does apply to all cards without restrictions.